오류 및 UI 수정사항 적용

+++ TowerDefense/Assets/01.Scripts/Tower/Base/Cores/Catapult_Core.cs
@@ -44,7 +44,7 @@ public class Catapult_Core : CoreBase
         }
     }
 
-    public override void Attack(int power, HealthSystem enemy)
+    public override void Attack(float power, HealthSystem enemy)
     {
         if (bullet == null && false == isThrowing)
         {
