+++ TowerDefense/Assets/01.Scripts/Tower/Base/Cores/Explosion_Core.cs
@@ -1,8 +1,6 @@
-using UnityEngine;
-
 public class Explosion_Core : CoreBase
 {
-    public override void Attack(int power, HealthSystem enemy)
+    public override void Attack(float power, HealthSystem enemy)
     {
         bullet = Managers.Pool.GetItem<Bomb>();
         bullet.transform.position = transform.position;
