+++ TowerDefense/Assets/01.Scripts/Tower/Base/Cores/Penetration_Core.cs
@@ -55,7 +55,7 @@ public class Penetration_Core : CoreBase
         }
     }
 
-    public override void Attack(int power, HealthSystem enemy)
+    public override void Attack(float power, HealthSystem enemy)
     {
         if (bullet == null && target != null)
         {
